+++ source_gen/source_gen/test/constants_test.dart
@@ -99,7 +99,7 @@ void main() {
 
       final nested = constant.read('nested');
       expect(nested.isNull, isFalse, reason: '$nested');
-      expect(nested.read('aString').isNull, isTrue);
+      expect(nested.read('aString').isNull, isTrue, reason: '$nested');
       expect(nested.read('aInt').isNull, isTrue);
       expect(nested.read('aBool').isNull, isTrue);
     });
@@ -148,6 +148,12 @@ void main() {
       final $null = constants[3];
       expect($null.isNull, isTrue, reason: '${$null}');
       expect(() => $null.read('foo'), throwsUnsupportedError);
+    });
+
+    test('should not fail reading from `null` when using peek', () {
+      final $null = constants[3];
+      expect($null.isNull, isTrue, reason: '${$null}');
+      expect($null.peek('foo'), isNull);
     });
 
     test('should fail reading a missing field', () {