+++ source_gen/source_gen/CHANGELOG.md
@@ -13,9 +13,19 @@
     `package:json_serializable`.
   * Removed `lib/builder.dart`. Import through `source_gen.dart` instead.
   * Removed `OutputFormatter` typedef.
+  
 * Add `LibraryReader.allElements` - a utility to iterate across all `Element`
   instances contained in Dart library.
 * Add `LibraryReader.element` to get back to the `LibraryElement` instance.
+* Add `ConstantReader.peek` to read a value that returns `null` if not found:
+
+```dart
+// Tries to read the field "token" first, then "_token".
+findTokenField(DartObject o) {
+  final reader = new ConstantReader(o);
+  final token = o.peek('token') ?? o.read('_token');
+}
+```
 
 ## 0.6.1+1